+++ src/main/java/net/frozenblock/wilderwild/worldgen/modification/WWSpawns.java
@@ -31,11 +31,8 @@ import net.minecraft.world.level.biome.Biomes;
 public final class WWSpawns {
 
 	public static void addBugs() {
-		BiomeModifications.addSpawn(BiomeSelectors.tag(WWBiomeTags.FIREFLY_SPAWNABLE_DURING_DAY),
-			FrozenMobCategories.getCategory(WWConstants.MOD_ID, "firefly"), WWEntityTypes.FIREFLY, 2, 12, 24);
-
-		BiomeModifications.addSpawn(BiomeSelectors.tag(WWBiomeTags.FIREFLY_SPAWNABLE),
-			FrozenMobCategories.getCategory(WWConstants.MOD_ID, "firefly"), WWEntityTypes.FIREFLY, 2, 12, 24);
+		BiomeModifications.addSpawn(BiomeSelectors.tag(WWBiomeTags.HAS_FIREFLY),
+			FrozenMobCategories.getCategory(WWConstants.MOD_ID, "firefly"), WWEntityTypes.FIREFLY, 4, 12, 24);
 
 		BiomeModifications.addSpawn(BiomeSelectors.tag(WWBiomeTags.HAS_BUTTERFLY),
 			FrozenMobCategories.getCategory(WWConstants.MOD_ID, "butterfly"), WWEntityTypes.BUTTERFLY, 1, 1, 1);